+++ GDX.Developer/Editor/Settings.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using System.IO;
+using GDX.Developer.Editor.Build;
 using GDX.Editor;
 using UnityEditor;
 using UnityEngine;
@@ -17,6 +19,13 @@ namespace GDX.Developer.Editor
         /// </summary>
         private static readonly HashSet<string> s_settingsKeywords =
             new HashSet<string>(new[] {"gdx", "developer", "parser"});
+
+        /// <summary>
+        ///     Settings content for <see cref="GDXConfig.developerBuildInfoAssemblyDefinition" />.
+        /// </summary>
+        private static readonly GUIContent s_contentBuildInfoAssemblyDefinition = new GUIContent(
+            "Assembly Definition",
+            "Ensure that the folder of the BuildInfo has an assembly definition.");
 
         /// <summary>
         ///     Settings content for <see cref="GDXConfig.developerCommandLineParserArgumentPrefix" />.
@@ -58,35 +67,35 @@ namespace GDX.Developer.Editor
         /// </summary>
         private static readonly GUIContent s_contentBuildInfoBuildNumberArgument = new GUIContent(
             "Number",
-            "The argument key for the build number to be passed to the BuildInfoGenerator.");
+            "The argument key for the build number to be passed to the BuildInfoProvider.");
 
         /// <summary>
         ///     Settings content for <see cref="GDXConfig.developerBuildInfoBuildDescriptionArgument" />.
         /// </summary>
         private static readonly GUIContent s_contentBuildInfoBuildDescriptionArgument = new GUIContent(
             "Description",
-            "The argument key for the build description to be passed to the BuildInfoGenerator.");
+            "The argument key for the build description to be passed to the BuildInfoProvider.");
 
         /// <summary>
         ///     Settings content for <see cref="GDXConfig.developerBuildInfoBuildChangelistArgument" />.
         /// </summary>
         private static readonly GUIContent s_contentBuildInfoBuildChangelistArgument = new GUIContent(
             "Changelist",
-            "The argument key for the build changelist to be passed to the BuildInfoGenerator.");
+            "The argument key for the build changelist to be passed to the BuildInfoProvider.");
 
         /// <summary>
         ///     Settings content for <see cref="GDXConfig.developerBuildInfoBuildTaskArgument" />.
         /// </summary>
         private static readonly GUIContent s_contentBuildInfoBuildTaskArgument = new GUIContent(
             "Task",
-            "The argument key for the build task to be passed to the BuildInfoGenerator.");
+            "The argument key for the build task to be passed to the BuildInfoProvider.");
 
         /// <summary>
         ///     Settings content for <see cref="GDXConfig.developerBuildInfoBuildStreamArgument" />.
         /// </summary>
         private static readonly GUIContent s_contentBuildInfoBuildStreamArgument = new GUIContent(
             "Stream",
-            "The argument key for the build stream to be passed to the BuildInfoGenerator.");
+            "The argument key for the build stream to be passed to the BuildInfoProvider.");
 
         /// <summary>
         ///     Get <see cref="SettingsProvider" /> for GDX.Developer assembly.
@@ -102,41 +111,48 @@ namespace GDX.Developer.Editor
                 guiHandler = searchContext =>
                 {
                     SerializedObject settings = ConfigProvider.GetSerializedConfig();
-                    SettingsGUILayout.BeginGUILayout();
-
-                    #region Command Line Parser
-
-                    SettingsGUILayout.SectionHeader("Command Line Parser");
-
-                    EditorGUILayout.PropertyField(settings.FindProperty("developerCommandLineParserArgumentPrefix"),
-                        s_contentArgumentPrefix);
-                    EditorGUILayout.PropertyField(settings.FindProperty("developerCommandLineParserArgumentSplit"),
-                        s_contentArgumentSplit);
-
-                    #endregion
-
-                    // Create some distance between next section
-                    GUILayout.Space(10);
+                    EditorGUILayout.BeginVertical(SettingsStyles.WrapperStyle);
 
                     #region BuildInfo Generation
 
-                    GUI.enabled = SettingsGUILayout.SectionHeader(
+                    bool buildInfoEnabled = SettingsStyles.SectionHeader(
                         "BuildInfo Generation",
                         settings.FindProperty("developerBuildInfoEnabled"),
                         s_contentBuildInfoEnabled);
 
+                    string buildInfoFile = Path.Combine(Application.dataPath,
+                        settings.FindProperty("developerBuildInfoPath").stringValue);
+                    if (!File.Exists(buildInfoFile))
+                    {
+                        GUILayout.BeginVertical(SettingsStyles.InfoBoxStyle);
+                        GUILayout.BeginHorizontal();
+                        GUILayout.Label(
+                            "There is currently no BuildInfo file in the target location\nWould you like some default content written in its place?");
+                        if (GUILayout.Button("Create Default", SettingsStyles.ButtonStyle))
+                        {
+                            BuildInfoProvider.WriteDefaultFile();
+                            AssetDatabase.ImportAsset("Assets/" +
+                                                      settings.FindProperty("developerBuildInfoPath").stringValue);
+                        }
+
+                        GUILayout.EndHorizontal();
+                        GUILayout.EndVertical();
+                    }
+
                     // Only allow editing based on the feature being enabled
-                    GUI.enabled = settings.FindProperty("developerBuildInfoEnabled").boolValue;
+                    GUI.enabled = buildInfoEnabled;
 
                     EditorGUILayout.PropertyField(settings.FindProperty("developerBuildInfoPath"),
                         s_contentBuildInfoPath);
                     EditorGUILayout.PropertyField(settings.FindProperty("developerBuildInfoNamespace"),
                         s_contentBuildInfoNamespace);
+                    EditorGUILayout.PropertyField(settings.FindProperty("developerBuildInfoAssemblyDefinition"),
+                        s_contentBuildInfoAssemblyDefinition);
 
 
                     GUILayout.Space(10);
                     // Arguments (we're going to make sure they are forced to uppercase).
-                    GUILayout.Label("Build Arguments", SettingsGUILayout.H2Style);
+                    GUILayout.Label("Build Arguments", SettingsStyles.SubSectionHeaderTextStyle);
 
                     SerializedProperty buildNumberProperty =
                         settings.FindProperty("developerBuildInfoBuildNumberArgument");
@@ -182,7 +198,21 @@ namespace GDX.Developer.Editor
 
                     #endregion
 
-                    SettingsGUILayout.EndGUILayout();
+                    // Create some distance between next section
+                    GUILayout.Space(10);
+
+                    #region Command Line Parser
+
+                    SettingsStyles.SectionHeader("Command Line Parser");
+
+                    EditorGUILayout.PropertyField(settings.FindProperty("developerCommandLineParserArgumentPrefix"),
+                        s_contentArgumentPrefix);
+                    EditorGUILayout.PropertyField(settings.FindProperty("developerCommandLineParserArgumentSplit"),
+                        s_contentArgumentSplit);
+
+                    #endregion
+
+                    EditorGUILayout.EndVertical();
                     settings.ApplyModifiedPropertiesWithoutUndo();
                 },
                 keywords = s_settingsKeywords