+++ GDX/Editor/Settings.cs
@@ -38,7 +38,6 @@ namespace GDX.Editor
             "Update Timer (Days)",
             "After how many days should updates be checked for?");
 
-
         /// <summary>
         ///     The number of days between checks for updates.
         /// </summary>
@@ -62,16 +61,16 @@ namespace GDX.Editor
                 guiHandler = searchContext =>
                 {
                     SerializedObject settings = ConfigProvider.GetSerializedConfig();
-                    SettingsGUILayout.BeginGUILayout();
+                    EditorGUILayout.BeginVertical(SettingsStyles.WrapperStyle);
 
                     #region Package Updates
 
-                    bool packageSectionEnabled = SettingsGUILayout.SectionHeader(
+                    bool packageSectionEnabled = SettingsStyles.SectionHeader(
                         "Automatic Package Updates",
                         settings.FindProperty("updateProviderCheckForUpdates"),
                         s_contentCheckForUpdates);
 
-                    EditorGUILayout.BeginHorizontal(SettingsGUILayout.InfoBoxStyle);
+                    EditorGUILayout.BeginHorizontal(SettingsStyles.InfoBoxStyle);
                     if (UpdateProvider.LocalPackage.Definition != null)
                     {
                         GUILayout.Label(
@@ -86,19 +85,19 @@ namespace GDX.Editor
                         EditorGUILayout.BeginVertical();
                         if (UpdateProvider.HasUpdate(UpdateProvider.UpdatePackageDefinition))
                         {
-                            if (GUILayout.Button("Changelog", SettingsGUILayout.ButtonStyle))
+                            if (GUILayout.Button("Changelog", SettingsStyles.ButtonStyle))
                             {
                                 Application.OpenURL(UpdateProvider.GitHubChangelogUri);
                             }
 
-                            if (GUILayout.Button("Update", SettingsGUILayout.ButtonStyle))
+                            if (GUILayout.Button("Update", SettingsStyles.ButtonStyle))
                             {
                                 UpdateProvider.AttemptUpgrade();
                             }
                         }
                         else
                         {
-                            if (GUILayout.Button("Manual Check", SettingsGUILayout.ButtonStyle))
+                            if (GUILayout.Button("Manual Check", SettingsStyles.ButtonStyle))
                             {
                                 UpdateProvider.CheckForUpdates();
                             }
@@ -126,8 +125,9 @@ namespace GDX.Editor
                     GUI.enabled = true;
 
                     #endregion
+
                     settings.ApplyModifiedPropertiesWithoutUndo();
-                    SettingsGUILayout.EndGUILayout();
+                    EditorGUILayout.EndVertical();
                 },
                 keywords = s_keywords
             };