+++ GDX/Editor/SettingsStyles.cs
@@ -4,6 +4,7 @@
 using UnityEditor;
 using UnityEngine;
 
+// ReSharper disable HeapView.ObjectAllocation.Evident
 // ReSharper disable MemberCanBePrivate.Global
 
 namespace GDX.Editor
@@ -11,93 +12,33 @@ namespace GDX.Editor
     /// <summary>
     ///     A helper class for generating the GDX editor experience.
     /// </summary>
-    public static class SettingsGUILayout
+    public static class SettingsStyles
     {
-        /// <summary>
-        ///     A shade of the <see cref="Color" /> green.
-        /// </summary>
-        /// <remarks>Meant for enabled things.</remarks>
-        private static readonly Color s_colorEnabledGreen =
-            new Color(0.1803921568627451f, 0.6431372549019608f, 0.3098039215686275f);
-
-        /// <summary>
-        ///     A shade of the <see cref="Color" /> blue.
-        /// </summary>
-        /// <remarks>Meant for default things.</remarks>
-        private static readonly Color s_colorDefaultBlue =
-            new Color(0.0941176470588235f, 0.4549019607843137f, 0.8549019607843137f);
-
-        /// <summary>
-        ///     A shade of the <see cref="Color" /> yellow.
-        /// </summary>
-        /// <remarks>Meant for disabled things.</remarks>
-        private static readonly Color s_colorDisabledYellow =
-            new Color(0.8941176470588235f, 0.9019607843137255f, 0.4117647058823529f);
-
-        /// <summary>
-        ///     A blendable shade of the <see cref="Color" /> white at 25% opacity.
-        /// </summary>
-        private static readonly Color s_colorWhiteBlend25 = new Color(1f, 1f, 1f, 0.25f);
-
-        /// <summary>
-        ///     A blendable shade of the <see cref="Color" /> white at 75% opacity.
-        /// </summary>
-        private static readonly Color s_colorWhiteBlend75 = new Color(1f, 1f, 1f, 0.75f);
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a horizontal line which respects margins/padding.
-        /// </summary>
-        private static readonly GUIStyle s_line;
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a default section header.
-        /// </summary>
-        private static readonly GUIStyle s_sectionHeader;
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a default section header text.
-        /// </summary>
-        private static readonly GUIStyle s_sectionHeaderTextDefault;
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a disabled section header text.
-        /// </summary>
-        private static readonly GUIStyle s_sectionHeaderTextDisabled;
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a enabled section header text.
-        /// </summary>
-        private static readonly GUIStyle s_sectionHeaderTextEnabled;
-
-        /// <summary>
-        ///     A collection of layout parameters to use when rendering the toggle option on section headers.
-        /// </summary>
-        private static readonly GUILayoutOption[] s_sectionHeaderToggleOptions;
-
-        /// <summary>
-        ///     A <see cref="GUIStyle" /> used to wrap all GDX editor user interfaces.
-        /// </summary>
-        private static readonly GUIStyle s_wrapper;
-
         /// <summary>
         ///     A <see cref="GUIStyle" /> representing a button.
         /// </summary>
         public static readonly GUIStyle ButtonStyle;
 
         /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a heading at level 1.
+        ///     A shade of the <see cref="Color" /> blue.
         /// </summary>
-        public static readonly GUIStyle H1Style;
+        /// <remarks>Meant for default things.</remarks>
+        public static readonly Color DefaultBlueColor =
+            new Color(0.0941176470588235f, 0.4549019607843137f, 0.8549019607843137f);
 
         /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a heading at level 2.
+        ///     A shade of the <see cref="Color" /> yellow.
         /// </summary>
-        public static readonly GUIStyle H2Style;
+        /// <remarks>Meant for disabled things.</remarks>
+        public static readonly Color DisabledYellowColor =
+            new Color(0.8941176470588235f, 0.9019607843137255f, 0.4117647058823529f);
 
         /// <summary>
-        ///     A <see cref="GUIStyle" /> representing a heading at level 3.
+        ///     A shade of the <see cref="Color" /> green.
         /// </summary>
-        public static readonly GUIStyle H3Style;
+        /// <remarks>Meant for enabled things.</remarks>
+        public static readonly Color EnabledGreenColor =
+            new Color(0.1803921568627451f, 0.6431372549019608f, 0.3098039215686275f);
 
         /// <summary>
         ///     A <see cref="GUIStyle" /> representing an info box.
@@ -105,61 +46,78 @@ namespace GDX.Editor
         public static readonly GUIStyle InfoBoxStyle;
 
         /// <summary>
-        ///     Initialize the <see cref="SettingsGUILayout" />, creating all of the associated <see cref="GUIStyle" />s.
+        ///     A <see cref="GUIStyle" /> representing a horizontal line which respects margins/padding.
         /// </summary>
-        static SettingsGUILayout()
+        public static readonly GUIStyle LineStyle;
+
+        /// <summary>
+        ///     A <see cref="GUIStyle" /> representing a default section header.
+        /// </summary>
+        public static readonly GUIStyle SectionHeaderStyle;
+
+        /// <summary>
+        ///     A <see cref="GUIStyle" /> representing a default section header text.
+        /// </summary>
+        public static readonly GUIStyle SectionHeaderTextDefaultStyle;
+
+        /// <summary>
+        ///     A <see cref="GUIStyle" /> representing a disabled section header text.
+        /// </summary>
+        public static readonly GUIStyle SectionHeaderTextDisabledStyle;
+
+        /// <summary>
+        ///     A collection of layout parameters to use when rendering the toggle option on section headers.
+        /// </summary>
+        public static readonly GUILayoutOption[] SectionHeaderToggleLayoutOptions;
+
+        /// <summary>
+        ///     A <see cref="GUIStyle" /> representing the header of a sub section definition.
+        /// </summary>
+        public static readonly GUIStyle SubSectionHeaderTextStyle;
+
+        /// <summary>
+        ///     A blendable shade of the <see cref="Color" /> white at 25% opacity.
+        /// </summary>
+        public static readonly Color WhiteBlend25Color = new Color(1f, 1f, 1f, 0.25f);
+
+        /// <summary>
+        ///     A blendable shade of the <see cref="Color" /> white at 75% opacity.
+        /// </summary>
+        public static readonly Color WhiteBlend75Color = new Color(1f, 1f, 1f, 0.75f);
+
+        /// <summary>
+        ///     A <see cref="GUIStyle" /> used to wrap all GDX editor user interfaces.
+        /// </summary>
+        public static readonly GUIStyle WrapperStyle;
+
+        /// <summary>
+        ///     Initialize the <see cref="SettingsStyles" />, creating all of the associated <see cref="GUIStyle" />s.
+        /// </summary>
+        static SettingsStyles()
         {
-            s_line =
+            LineStyle =
                 new GUIStyle("box")
                 {
                     border = {top = 0, bottom = 0},
                     margin = {top = 0, bottom = 0, left = 0, right = 0}, // makes sure to
                     padding = {top = 0, bottom = 0}
                 };
-
             InfoBoxStyle = new GUIStyle(EditorStyles.helpBox)
             {
                 padding = {top = 10, bottom = 10, left = 10, right = 10}, margin = {bottom = 10}
             };
-
-            s_wrapper = new GUIStyle {margin = {left = 5, right = 5, bottom = 5}};
-
-
+            WrapperStyle = new GUIStyle {margin = {left = 5, right = 5, bottom = 5}};
             ButtonStyle =
                 new GUIStyle(EditorStyles.miniButton) {stretchWidth = false, padding = {left = 10, right = 10}};
 
-            // Build our headers
-            H1Style = new GUIStyle(EditorStyles.largeLabel) {fontStyle = FontStyle.Bold};
-            H2Style = new GUIStyle(H1Style) {fontSize = H1Style.fontSize - 1, margin = { left = 2 }};
-            H3Style = new GUIStyle(H2Style) {fontSize = H2Style.fontSize - 1};
-
             // Section Headers
-            s_sectionHeader = new GUIStyle("box") {margin = {left = -20}};
-            s_sectionHeaderTextDefault = new GUIStyle(H1Style);
-            s_sectionHeaderTextEnabled =
-                new GUIStyle(s_sectionHeaderTextDefault) {normal = {textColor = s_colorWhiteBlend75}};
-            s_sectionHeaderTextDisabled =
-                new GUIStyle(s_sectionHeaderTextDefault) {normal = {textColor = s_colorWhiteBlend25}};
-
-            s_sectionHeaderToggleOptions = new[] {GUILayout.Width(EditorStyles.toggle.CalcSize(GUIContent.none).x)};
+            SectionHeaderStyle = new GUIStyle("box") {margin = {left = -20}};
+            SectionHeaderTextDefaultStyle = new GUIStyle(EditorStyles.largeLabel) {fontStyle = FontStyle.Bold, normal = {textColor = WhiteBlend75Color}};
+            SectionHeaderTextDisabledStyle =
+                new GUIStyle(SectionHeaderTextDefaultStyle) {normal = {textColor = WhiteBlend25Color}};
+            SectionHeaderToggleLayoutOptions = new[] {GUILayout.Width(EditorStyles.toggle.CalcSize(GUIContent.none).x)};
+            SubSectionHeaderTextStyle = new GUIStyle(EditorStyles.largeLabel) {fontStyle = FontStyle.Bold, fontSize = EditorStyles.largeLabel.fontSize - 1, margin = { left = 2 }};
         }
-
-        /// <summary>
-        ///     Simplified start of a wrapped user interface experience.
-        /// </summary>
-        public static void BeginGUILayout()
-        {
-            EditorGUILayout.BeginVertical(s_wrapper);
-        }
-
-        /// <summary>
-        ///     Simplified end of a wrapped user interface experience.
-        /// </summary>
-        public static void EndGUILayout()
-        {
-            EditorGUILayout.EndVertical();
-        }
-
 
         /// <summary>
         ///     Draw a line during the definition of a user interface experience which respects padding/margins, but also adds its
@@ -178,7 +136,7 @@ namespace GDX.Editor
             }
 
             // Draw line of sorts
-            GUILayout.Box(GUIContent.none, s_line,
+            GUILayout.Box(GUIContent.none, LineStyle,
                 GUILayout.ExpandWidth(true),
                 GUILayout.Height(height));
 
@@ -206,9 +164,9 @@ namespace GDX.Editor
 
             if (sectionToggleProperty == null)
             {
-                GUI.backgroundColor = s_colorDefaultBlue;
-                GUILayout.BeginHorizontal(s_sectionHeader);
-                GUILayout.Label(text, s_sectionHeaderTextDefault);
+                GUI.backgroundColor = DefaultBlueColor;
+                GUILayout.BeginHorizontal(SectionHeaderStyle);
+                GUILayout.Label(text, SectionHeaderTextDefaultStyle);
                 GUILayout.EndHorizontal();
                 GUI.backgroundColor = previousColor;
                 GUILayout.Space(5);
@@ -217,21 +175,21 @@ namespace GDX.Editor
 
             if (sectionToggleProperty.boolValue)
             {
-                GUI.backgroundColor = s_colorEnabledGreen;
-                GUILayout.BeginHorizontal(s_sectionHeader);
-                GUILayout.Label(text, s_sectionHeaderTextEnabled);
+                GUI.backgroundColor = EnabledGreenColor;
+                GUILayout.BeginHorizontal(SectionHeaderStyle);
+                GUILayout.Label(text, SectionHeaderTextDefaultStyle);
                 GUILayout.FlexibleSpace();
                 EditorGUILayout.PropertyField(sectionToggleProperty, sectionToggleContent,
-                    s_sectionHeaderToggleOptions);
+                    SectionHeaderToggleLayoutOptions);
             }
             else
             {
-                GUI.backgroundColor = s_colorDisabledYellow;
-                GUILayout.BeginHorizontal(s_sectionHeader);
-                GUILayout.Label(text, s_sectionHeaderTextDisabled);
+                GUI.backgroundColor = DisabledYellowColor;
+                GUILayout.BeginHorizontal(SectionHeaderStyle);
+                GUILayout.Label(text, SectionHeaderTextDisabledStyle);
                 GUILayout.FlexibleSpace();
                 EditorGUILayout.PropertyField(sectionToggleProperty, sectionToggleContent,
-                    s_sectionHeaderToggleOptions);
+                    SectionHeaderToggleLayoutOptions);
             }
 
             GUILayout.EndHorizontal();